Update 1D-avection code and test cases

+++ src/c++/1d/1d_advection/packages/src/adv_diff_source.cpp
@@ -21,7 +21,7 @@ void adv_diff_source(std::vector<double> & q, int select)
         //
         // Homogeneous righthand side
         //
-        for (int i = 0; i < q.size(); ++i)
+        for (size_t i = 0; i < q.size(); ++i)
         {
             q[i] = 0.0;
         }
